move alertingSchemaApi.ts and the alerting section in README.md to a new alerting folder
